+++ packages/core/src/index.ts
@@ -40,3 +40,4 @@ export * as karmaStreaks from "./karma_streaks.js";
 export * as nucleus from "./nucleus.js";
 export * as nucleusDaemon from "./nucleus_daemon.js";
 export * as lineageSeed from "./lineage_seed.js";
+export * as inbox from "./inbox.js";